+++ skops/_min_dependencies.py
@@ -34,6 +34,7 @@ dependent_packages = {
     # TODO: remove condition when catboost supports python 3.11
     "catboost": ("1.0", "tests", "python_version < '3.11'"),
     "fairlearn": ("0.7.0", "docs, tests", None),
+    "rich": ("12", "tests", None),
 }