+++ skops/io/tests/test_visualize.py
@@ -1,3 +1,7 @@
+"""Tests for skops.io.visualize"""
+
+from unittest.mock import Mock, patch
+
 import pytest
 import sklearn
 from sklearn.linear_model import LogisticRegression
@@ -10,7 +14,6 @@ from sklearn.preprocessing import (
 )
 
 import skops.io as sio
-from skops.io._visualize import visualize_tree
 
 
 class TestVisualizeTree:
@@ -44,10 +47,10 @@ class TestVisualizeTree:
     @pytest.mark.parametrize("show", ["all", "trusted", "untrusted"])
     def test_print_simple(self, simple, show, capsys):
         file = sio.dumps(simple)
-        visualize_tree(file, show=show)
+        sio.visualize_tree(file, show=show)
 
-        # output is always the same for "all" and "trusted" because all nodes
-        # are trusted
+        # Output is the same for "all" and "trusted" because all nodes are
+        # trusted. Colors are not recorded by capsys.
         expected = [
             "root: sklearn.preprocessing._data.MinMaxScaler",
             "└── attrs: builtins.dict",
@@ -56,7 +59,7 @@ class TestVisualizeTree:
             "    │   └── content: json-type(123)",
             "    ├── copy: json-type(true)",
             "    ├── clip: json-type(false)",
-            f'    └── _sklearn_version: json-type("{sklearn.__version__}")',
+            '    └── _sklearn_version: json-type("{}")'.format(sklearn.__version__),
         ]
         if show == "untrusted":
             # since no untrusted, only show root
@@ -65,9 +68,9 @@ class TestVisualizeTree:
         stdout, _ = capsys.readouterr()
         assert stdout.strip() == "\n".join(expected)
 
-    def test_print_pipelien(self, pipeline, capsys):
+    def test_print_pipeline(self, pipeline, capsys):
         file = sio.dumps(pipeline)
-        visualize_tree(file)
+        sio.visualize_tree(file)
 
         # no point in checking the whole output with > 120 lines
         expected_start = [
@@ -80,7 +83,7 @@ class TestVisualizeTree:
         expected_end = [
             "    ├── memory: json-type(null)",
             "    ├── verbose: json-type(false)",
-            f'    └── _sklearn_version: json-type("{sklearn.__version__}")',
+            '    └── _sklearn_version: json-type("{}")'.format(sklearn.__version__),
         ]
 
         stdout, _ = capsys.readouterr()
@@ -94,7 +97,7 @@ class TestVisualizeTree:
         def sink(nodes_iter, *args, **kwargs):
             nodes.extend(nodes_iter)
 
-        visualize_tree(file, sink=sink)
+        sio.visualize_tree(file, sink=sink)
         nodes_self_unsafe = [node for node in nodes if not node.is_self_safe]
         nodes_unsafe = [node for node in nodes if not node.is_safe]
 
@@ -111,11 +114,88 @@ class TestVisualizeTree:
         assert len(nodes_unsafe) > 2
         assert any("FunctionTransformer" in node.val for node in nodes_unsafe)
 
-    def test_custom_print_config(self):
-        pass
+    @pytest.mark.parametrize(
+        "kwargs",
+        [
+            {},
+            {"use_colors": False},
+            {"tag_unsafe": "<careful>", "color_unsafe": "blue"},
+        ],
+    )
+    def test_custom_print_config_passed_to_sink(self, simple, kwargs):
+        # check that arguments are passed to sink
+        def my_sink(nodes_iter, show, **sink_kwargs):
+            for key, val in kwargs.items():
+                assert sink_kwargs[key] == val
 
-    def test_from_file(self):
-        pass
+        file = sio.dumps(simple)
+        sio.visualize_tree(file, sink=my_sink, **kwargs)
 
-    def test_rich_not_installed(self):
-        pass
+    def test_custom_tags(self, simple, capsys):
+        class UnsafeType:
+            pass
+
+        simple.copy = UnsafeType
+
+        file = sio.dumps(simple)
+        sio.visualize_tree(file, tag_safe="NICE", tag_unsafe="OHNO")
+        expected = [
+            "root: sklearn.preprocessing._data.MinMaxScaler NICE",
+            "└── attrs: builtins.dict NICE",
+            "    ├── feature_range: builtins.tuple NICE",
+            "    │   ├── content: json-type(-555) NICE",
+            "    │   └── content: json-type(123) NICE",
+            "    ├── copy: test_visualize.UnsafeType OHNO",
+            "    ├── clip: json-type(false) NICE",
+            '    └── _sklearn_version: json-type("{}") NICE'.format(
+                sklearn.__version__
+            ),
+        ]
+
+        stdout, _ = capsys.readouterr()
+        assert stdout.strip() == "\n".join(expected)
+
+    def test_custom_colors(self, simple):
+        # Colors are not recorded by capsys, so we cannot use it
+        class UnsafeType:
+            pass
+
+        simple.copy = UnsafeType
+
+        file = sio.dumps(simple)
+        mock_print = Mock()
+        with patch("rich.print", mock_print):
+            sio.visualize_tree(
+                file,
+                color_safe="black",
+                color_unsafe="cyan",
+                color_child_unsafe="orange3",
+            )
+
+        assert mock_print.call_count == 1
+
+        tree = mock_print.call_args_list[0].args[0]
+        # The root node is indirectly unsafe through child
+        assert "[orange3]" in tree.label
+        # feature_range is safe
+        assert "[black]" in tree.children[0].children[0].label
+        # copy is unsafe
+        assert "[cyan]" in tree.children[0].children[1].label
+
+    def test_from_file(self, simple, tmp_path, capsys):
+        f_name = tmp_path / "estimator.skops"
+        sio.dump(simple, f_name)
+        sio.visualize_tree(f_name)
+
+        expected = [
+            "root: sklearn.preprocessing._data.MinMaxScaler",
+            "└── attrs: builtins.dict",
+            "    ├── feature_range: builtins.tuple",
+            "    │   ├── content: json-type(-555)",
+            "    │   └── content: json-type(123)",
+            "    ├── copy: json-type(true)",
+            "    ├── clip: json-type(false)",
+            '    └── _sklearn_version: json-type("{}")'.format(sklearn.__version__),
+        ]
+        stdout, _ = capsys.readouterr()
+        assert stdout.strip() == "\n".join(expected)