+++ docs/changes.rst
@@ -20,6 +20,9 @@ v0.6
   `Benjamin Bossan`_.
 - Fix: skops persistence now also works with many functions from the
   :mod:`operator` module. :pr:`287` by `Benjamin Bossan`_.
+- Add possibility to visualize a skops object and show untrusted types by using
+  :func:`skops.io.visualize_tree`. Requires to install `rich`: `pip install
+  rich`. :pr:`317` by `Benjamin Bossan`_.
 
 v0.5
 ----